+++ website/playground/codeSamples/css.txt
@@ -1,0 +1,10 @@
+@media (max-width: 480px) {
+  .bd-examples {margin-right: -.75rem;margin-left: -.75rem
+  }
+  
+ .bd-examples>[class^="col-"]  {
+    padding-right: .75rem;
+    padding-left: .75rem;
+  
+  }
+}
