+++ website/playground/codeSamples/flow.txt
@@ -1,0 +1,8 @@
+declare export function graphql<Props, Variables, Component: React$ComponentType<Props>>
+  (query: GQLDocument, config?: Config<Props, QueryConfigOptions<Variables>>):
+  (Component: Component) => React$ComponentType<$Diff<React$ElementConfig<Component>, {
+    data: Object|void,
+    mutate: Function|void
+  }>>
+
+declare type FetchPolicy = "cache-first" | "cache-and-network" | "network-only" | "cache-only"
