+++ website/playground/codeSamples/glimmer.txt
@@ -1,0 +1,5 @@
+  <div     class="entry"    >
+  <h1>{{  title    }}</h1>
+  <div   class="body">
+            {{   body         }}
+</div> </div>
