+++ website/playground/codeSamples/graphql.txt
@@ -1,0 +1,9 @@
+query Browse($offset: Int, $limit: Int, $categories: [String!], $search: String) {
+  browse(limit: $limit, offset: $offset, categories: $categories, search: $search) {
+    total,
+    results {
+        title
+        price
+    }
+  }
+}
