+++ website/playground/codeSamples/json.txt
@@ -1,0 +1,13 @@
+{"allOn": "Single", "Line": "example",
+"noSpace":true,
+  "quote": {
+    'singleQuote': 'example',
+                  "indented": true,
+  },
+  "phoneNumbers": [
+    {"type": "home",
+      "number": "212 555-1234"},
+    {"type": "office",
+      "trailing": "commas by accident"},
+  ],
+}
