+++ website/playground/codeSamples/less.txt
@@ -1,0 +1,10 @@
+@my-ruleset: {
+    .my-selector {
+      @media tv {
+        background-color: black;
+      }
+    }
+  };
+@media (orientation:portrait) {
+    @my-ruleset();
+}
