+++ website/playground/codeSamples/markdown.txt
@@ -1,0 +1,25 @@
+Header
+======
+
+_Look,_ code blocks are formatted *too!*
+
+``` js
+function identity(x) { return x }
+```
+
+Pilot|Airport|Hours
+--|:--:|--:
+John Doe|SKG|1338
+Jane Roe|JFK|314
+
+- - - - - - - - - - - - - - -
+
++ List
+ + with a [link] (/to/somewhere)
++ and [another one]
+
+
+  [another one]:  http://example.com 'Example title'
+
+Lorem ipsum dolor sit amet, consectetur adipiscing elit.
+Curabitur consectetur maximus risus, sed maximus tellus tincidunt et.
