+++ website/playground/codeSamples/mdx.txt
@@ -1,0 +1,41 @@
+import     {     Baz } from     './Fixture'
+import { Buz  }   from './Fixture'
+
+export  const   foo    = {
+  hi:     `Fudge ${Baz.displayName || 'Baz'}`,
+  authors: [
+     'fred',
+           'sally'
+    ]
+}
+
+# Hello,    world!
+
+
+ I'm an awesome   paragraph.
+
+<!-- I'm a comment -->
+
+<Foo bg='red'>
+      <Bar    >hi    </Bar>
+       {  hello       }
+       {     /* another comment */}
+</Foo>
+
+```
+test codeblock
+```
+
+```js
+module.exports = 'test'
+```
+
+```sh
+npm i -g foo
+```
+
+| Test  | Table   |
+|    :---     | :----  |
+|   Col1  | Col2    |
+
+export   default     ({children   }) => < div>{    children}</div>
