+++ website/playground/codeSamples/vue.txt
@@ -1,0 +1,12 @@
+<template>
+  <p>Templates are formatted as well...
+    </p>
+</template>
+
+<script>
+let Prettier = format => { your.js('though') }
+</script>
+
+<style>
+.and { css: too !important }
+</style>
